pyproject.toml
[project]
name = "visual-grounding"
version = "0.1.0"
description = "Research-ready visual grounding pipelines with TransVG backbones"
readme = "README.md"
authors = [
    { name = "Viet Nguyen", email = "vietnguyen.nq@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7",
    "lmdb>=1.4",
    "matplotlib>=3.8",
    "numpy>=1.24",
    "pillow>=10.0",
    "timm>=0.9",
    "torch>=2.1",
    "torchvision>=0.16",
    "tqdm>=4.66",
    "transformers>=4.39",
]

[project.scripts]
visual-grounding = "visual_grounding.__main__:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
